Added the Files used for Token Retireval
Ebay Authentication uses oAuth. This code is what I used to generate an Authentication token for a eBay developer account for the use of APIs

**EBAY USES HTTPS FOR THEIR SERVERS WITH oAUTH. YOU MUST HOST YOUR LOCALHOST SERVER ON A HTTPS SERVER**
this is why if you look at my code that creates the express server at the bottom of the app.js file it looks weird.